backend/pyproject.toml
[project]
name = "anki-deck-builder"
version = "0.1.0"
description = "Anki deck builder with OpenAI integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openai>=1.12.0",
    "genanki>=0.13.1",
    "python-dotenv>=1.0.0",
    "stripe>=12.1.0",
    "pydantic-settings>=2.9.1",
    "google>=3.0.0",
    "google-cloud-storage>=3.1.0",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.23.0",
]
